--- a/include/fastscapelib/utils.hpp
+++ b/include/fastscapelib/utils.hpp
@@ -2,7 +2,9 @@
  * @file
  * @brief Some utilities used internally.
  */
-#pragma once
+#ifndef FASTSCAPELIB_UTILS_H
+#define FASTSCAPELIB_UTILS_H
+
 
 #include <cstddef>
 #include <type_traits>
@@ -25,27 +27,23 @@ using xtensor_t = xt::xexpression<E>;
 
 namespace fastscapelib
 {
-
-namespace detail
-{
-
-
-/**
- * @brief Return true if a given (row, col) index is in bounds (false
- *        otherwise).
- */
-template<class S>
-bool in_bounds(const S& shape, index_t row, index_t col)
-{
-    if(row >= 0 && row < static_cast<index_t>(shape[0])
-       && col >= 0 && col < static_cast<index_t>(shape[1]))
+    namespace detail
     {
-        return true;
+        /**
+         * @brief Return true if a given (row, col) index is in bounds (false
+         *        otherwise).
+         */
+        template<class S>
+        bool in_bounds(const S& shape, index_t row, index_t col)
+        {
+            if(row >= 0 && row < static_cast<index_t>(shape[0])
+            && col >= 0 && col < static_cast<index_t>(shape[1]))
+            {
+                return true;
+            }
+            return false;
+        }
     }
-    return false;
 }
 
-
-}  // namespace detail
-
-}  // namespace fastscapelib
+#endif