This PR is a major refactoring of `raster_grid` class. It is related to #61.

The `raster_connect` is now a template parameter, which allow optimization of the neighbors indices cache based on fixed size objects.

- [x] Rebase after #60 merged

#### Different `raster_connect` than the original one
Capability to easily access neighbors from the class `raster_connect` to an other is preserved using the newly introduced `create_similar_raster` template method:
```cpp
using grid_type = fs::raster_grid_xt<fs::xtensor_selector, fs::raster_connect::queen>;
using shape_type = typename grid_type::shape_type;

shape_type shape {{5, 10}};
grid_type queen_grid = grid_type(shape, {1.3, 1.2}, node_status::fixed_value_boundary);

auto rook_grid = queen_grid.create_similar_raster<raster_connect::rook>();
```

#### Benchmarks modifications
Improved readability of benchmarks:
- reduce number of tested size, but increase configurations (grids, raster_connect, cache policies)
- use macros to reduce boilerplate
- use `using` declarations to make benchs outputs nicer

ex:
*before*
```
flow_routing__compute_receivers__profile<fs::profile_grid>/256                        2 us          2 us     388090
flow_routing__compute_receivers__profile<fs::profile_grid>/512                        4 us          4 us     179656
flow_routing__compute_receivers__profile<fs::profile_grid>/1024                       9 us          9 us      72550
flow_routing__compute_receivers__profile<fs::profile_grid>/2048                      23 us         23 us      30516
flow_routing__compute_receivers__profile<fs::profile_grid>/4096                      52 us         52 us      12755
flow_routing__compute_receivers__raster<fs::raster_grid>/256/min_time:2.000           2 ms          2 ms       1199
flow_routing__compute_receivers__raster<fs::raster_grid>/512/min_time:2.000          10 ms         10 ms        294
flow_routing__compute_receivers__raster<fs::raster_grid>/1024/min_time:2.000         38 ms         38 ms         73
flow_routing__compute_receivers__raster<fs::raster_grid>/2048/min_time:2.000        167 ms        167 ms         17
flow_routing__compute_receivers__raster<fs::raster_grid>/4096/min_time:2.000        664 ms        664 ms          4
flow_routing__compute_receivers_d8/256/min_time:2.000                                 3 ms          3 ms        938
flow_routing__compute_receivers_d8/512/min_time:2.000                                12 ms         12 ms        241
flow_routing__compute_receivers_d8/1024/min_time:2.000                               46 ms         46 ms         60
flow_routing__compute_receivers_d8/2048/min_time:2.000                              205 ms        204 ms         10
flow_routing__compute_receivers_d8/4096/min_time:2.000                              812 ms        809 ms          4
```
*after*
```
flow_routing__compute_receivers__profile<profile_nocache>/1024          12 us         12 us      58783
flow_routing__compute_receivers__profile<profile_cacheall>/1024         10 us         10 us      69031
flow_routing__compute_receivers__raster<queen_nocache>/1024             53 ms         53 ms         13
flow_routing__compute_receivers__raster<queen_cacheall>/1024            40 ms         40 ms         17
flow_routing__compute_receivers__raster<rook_nocache>/1024              35 ms         35 ms         21
flow_routing__compute_receivers__raster<rook_cacheall>/1024             27 ms         27 ms         26
flow_routing__compute_receivers__raster<bishop_nocache>/1024            39 ms         39 ms         18
flow_routing__compute_receivers__raster<bishop_cacheall>/1024           29 ms         29 ms         25
flow_routing__compute_receivers_d8/1024                                 47 ms         47 ms         15
```

#### Details

This PR includes:
- used template parameter for raster_connect
- inherit raster_connects specific props and methods
- implement bishop `raster_connect`
- extend testing and benchmarking
- simplify bench, focus on public API