@benbovy 2021-01-12
Nice!

> Capability to easily access neighbors from the class raster_connect to an other is preserved using the newly introduced create_similar_raster template method

I like the idea. That's one more step than, e.g., `queen_grid.neighbors<raster_connect::rook>(idx);`, but I guess the implementation is much easier with another grid object and no template parameter for `.neighbors()`?

In the current implementation, does `queen_grid.create_similar_raster<raster_connect::rook>();` reuse the cache from from `queen_grid`. If not, my concern is that it may waste too much memory for larger grids.

Nit: I'm not super fan of the name `create_similar_raster`. Depending on whether or not it reuses the grid cache, an alternative suggestion might be `clone_grid`. Naive question: is it possible to have a static factory method instead?