@adriendelsalle 2021-01-12
> but I guess the implementation is much easier with another grid object and no template parameter for `.neighbors()`

Yes it is!

> In the current implementation, does queen_grid.create_similar_raster<raster_connect::rook>(); reuse the cache from from queen_grid? If not, my concern is that it may waste too much memory for larger grids.

I was also thinking about create a `view` on the cache if already filled and if it is a superset (e.g. `rook` and `bishop` can use the `queen` cache, but a `queen` cannot reuse the cache of a `rook`), but it makes the code really complicated and not efficient because of the filtering vs allocation and recomputation of the cache.
You're right about the memory footprint and we probably need to work more on it !

> Nit: I'm not super fan of the name create_similar_raster. Depending on whether or not it reuses the grid cache, an alternative suggestion might be clone_grid. Naive question: is it possible to have a static factory method instead?

No problem for the renaming! I will also try to make it static to follow a factory pattern