@adriendelsalle 2021-01-12
@benbovy 

Do you prefer the following semantic?
```cpp
using queen_type = fs::raster_grid_xt<fs::xtensor_selector, fs::raster_connect::queen>;

shape_type shape {{5, 10}};
grid_type queen_grid = grid_type(shape, {1.3, 1.2}, node_status::fixed_value_boundary);

auto rook_like = fs::raster_grid::clone_grid<raster_connect::rook>(queen_fixed);
```

Maybe `clone` is enough ?

Thanks for your feedback !