@benbovy 2021-01-12
> Do you prefer the following semantic?

Yes I like it. Actually, I'm fine with either a factory or a instance method :-)

> Maybe clone is enough ?

Agreed!