@adriendelsalle 2021-01-13
Updated
--
- benchmarks only reduced for `grids` (keep 2 sizes)
- `create_similar_raster` renamed `clone` and made static
- #66 opened to deal with neighbors indices cache reuse